requirements.txt
# GradeIt Dependencies

# AI Integration
anthropic>=0.18.0
google-generativeai>=0.3.0

# Git operations
gitpython>=3.1.40

# Configuration parsing
python-dotenv>=1.0.0

# CLI utilities
click>=8.1.7
tqdm>=4.66.0

# File operations and utilities
pathlib>=1.0.1
